[#422] refactor : editPage 코드 위치 변경
